# apps/api/docs/openapi.yaml
components:
  schemas:
...
    CommunicationQueue:
      description: Priority communication item in the queue.
      properties:
        acknowledged:
          default: false
          title: Acknowledged
          type: boolean
        acknowledged_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Acknowledged At
        attempts:
          default: 0
          minimum: 0.0
          title: Attempts
          type: integer
        content:
          title: Content
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        delivered:
          default: false
          title: Delivered
          type: boolean
        escalation_level:
          default: 0
          description: 0=nudge, 1=recommendation, 2=order
          maximum: 2.0
          minimum: 0.0
          title: Escalation Level
          type: integer
        id:
          title: Id
          type: string
        metadata:
          title: Metadata
          type: object
        priority:
          default: 5
          description: Priority 1-10, higher = more important
          maximum: 10.0
          minimum: 1.0
          title: Priority
          type: integer
        related_goals:
          items:
            type: string
          title: Related Goals
          type: array
        related_truths:
          items:
            type: string
          title: Related Truths
          type: array
        scheduled_for:
          format: date-time
          title: Scheduled For
          type: string
        source_report:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Report
        tenant_id:
          title: Tenant Id
          type: string
        topic:
          title: Topic
          type: string
        type:
          allOf:
          - $ref: '#/components/schemas/CommunicationType'
          default: nudge
        user_id:
          title: User Id
          type: string
      required:
      - user_id
      - tenant_id
      - topic
      - content
      title: CommunicationQueue
      type: object
    CommunicationType:
      description: Types of priority communications.
      enum:
      - nudge
      - recommendation
      - order
      title: CommunicationType
      type: string
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ImpactLevel:
      description: Impact levels for strategic insights.
      enum:
      - low
      - medium
      - high
      - critical
      title: ImpactLevel
      type: string
...
    StrategicAlignmentScorecard:
      description: Comprehensive strategic alignment scorecard for an organization.
      properties:
        adaptation_speed:
          default: 0.0
          maximum: 100.0
          minimum: 0.0
          title: Adaptation Speed
          type: number
        alignment_zone:
          allOf:
          - $ref: '#/components/schemas/StrategicAlignmentZone'
          default: red
        cross_functional_alignment:
          default: 0.0
          maximum: 100.0
          minimum: 0.0
          title: Cross Functional Alignment
          type: number
        decision_strategy_consistency:
          default: 0.0
          maximum: 100.0
          minimum: 0.0
          title: Decision Strategy Consistency
          type: number
        goal_cascade_alignment:
          default: 0.0
          maximum: 100.0
          minimum: 0.0
          title: Goal Cascade Alignment
          type: number
        id:
          title: Id
          type: string
        measurement_date:
          format: date-time
          title: Measurement Date
          type: string
        metadata:
          title: Metadata
          type: object
        overall_alignment_score:
          default: 0.0
          maximum: 100.0
          minimum: 0.0
          title: Overall Alignment Score
          type: number
        priority_interventions:
          items:
            type: string
          title: Priority Interventions
          type: array
        resource_allocation_efficiency:
          default: 0.0
          maximum: 100.0
          minimum: 0.0
          title: Resource Allocation Efficiency
          type: number
        risk_indicators:
          items:
            type: string
          title: Risk Indicators
          type: array
        strategic_communication_effectiveness:
          default: 0.0
          maximum: 100.0
          minimum: 0.0
          title: Strategic Communication Effectiveness
          type: number
        strategic_initiative_velocity:
          default: 0.0
          maximum: 100.0
          minimum: 0.0
          title: Strategic Initiative Velocity
          type: number
        strategic_response_time:
          default: 0.0
          maximum: 100.0
          minimum: 0.0
          title: Strategic Response Time
          type: number
        strategic_velocity:
          default: 0.0
          maximum: 100.0
          minimum: 0.0
          title: Strategic Velocity
          type: number
        tenant_id:
          title: Tenant Id
          type: string
        trend_30_days:
          default: stable
          description: improving, declining, stable
          title: Trend 30 Days
          type: string
        trend_60_days:
          default: stable
          description: improving, declining, stable
          title: Trend 60 Days
          type: string
        trend_90_days:
          default: stable
          description: improving, declining, stable
          title: Trend 90 Days
          type: string
      required:
      - tenant_id
      title: StrategicAlignmentScorecard
      type: object
    StrategicAlignmentZone:
      description: Strategic alignment health zones.
      enum:
      - red
      - yellow
      - green
      title: StrategicAlignmentZone
      type: string
...
    TruthCategory:
      description: Categories for strategic insights.
      enum:
      - technology
      - market
      - competitive
      - regulatory
      - social
      - operational
      title: TruthCategory
      type: string
...
info:
  contact:
    email: support@livingtwin.com
    name: Living Twin API Support
    url: https://github.com/kpernyer/living-twin-monorep
  description: "\n# Living Twin Strategic Intelligence Platform\n\nA comprehensive\
    \ API for strategic intelligence, competitive analysis, and AI-powered business\
    \ insights.\n\n## Key Capabilities\n\n* **Strategic Intelligence**: SWOT analysis,\
    \ Porter's Five Forces, competitive intelligence\n* **RAG System**: Document ingestion,\
    \ semantic search, conversational AI\n* **Health Monitoring**: Comprehensive system\
    \ health and performance monitoring  \n* **Testing Framework**: End-to-end testing\
    \ for strategic intelligence workflows\n\n## Authentication\n\nAll endpoints require\
    \ Firebase authentication via the `Authorization` header:\n```\nAuthorization:\
    \ Bearer <firebase-jwt-token>\n```\n\nDevelopment environments support bypass\
    \ mode for testing.\n\n## Tenant Isolation\n\nAll data is isolated by tenant.\
    \ Cross-tenant access is controlled by role-based permissions.\n"
  license:
    name: Proprietary
    url: https://livingtwin.com/license
  title: Living Twin API
  version: 2.0.0
openapi: 3.1.0
paths:
...
  /intelligence/alignment/history:
    get:
      description: Get historical strategic alignment data.
      operationId: get_strategic_alignment_history_intelligence_alignment_history_get
      parameters:
      - in: query
        name: date_from
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Date From
      - in: query
        name: date_to
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Date To
      - in: query
        name: limit
        required: false
        schema:
          default: 30
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/StrategicAlignmentScorecard'
                title: Response Get Strategic Alignment History Intelligence Alignment
                  History Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Strategic Alignment History
      tags:
      - intelligence
  /intelligence/alignment/scorecard:
    get:
      description: Get strategic alignment scorecard for the current tenant.
      operationId: get_strategic_alignment_scorecard_intelligence_alignment_scorecard_get
      parameters:
      - in: query
        name: date_from
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Date From
      - in: query
        name: date_to
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Date To
      - in: query
        name: include_details
        required: false
        schema:
          default: false
          title: Include Details
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategicAlignmentScorecard'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Strategic Alignment Scorecard
      tags:
      - intelligence
  /intelligence/communications:
    get:
      description: Get priority communications for the current user.
      operationId: get_priority_communications_intelligence_communications_get
      parameters:
      - in: query
        name: types
        required: false
        schema:
          anyOf:
          - items:
              $ref: '#/components/schemas/CommunicationType'
            type: array
          - type: 'null'
          title: Types
      - in: query
        name: priority_min
        required: false
        schema:
          anyOf:
          - maximum: 10
            minimum: 1
            type: integer
          - type: 'null'
          title: Priority Min
      - in: query
        name: delivered
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Delivered
      - in: query
        name: acknowledged
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Acknowledged
      - in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CommunicationQueue'
                title: Response Get Priority Communications Intelligence Communications
                  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Priority Communications
      tags:
      - intelligence
  /intelligence/communications/{communication_id}/acknowledge:
    post:
      description: Acknowledge a priority communication.
      operationId: acknowledge_priority_communication_intelligence_communications__communication_id__acknowledge_post
      parameters:
      - in: path
        name: communication_id
        required: true
        schema:
          title: Communication Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Acknowledge Priority Communication
      tags:
      - intelligence
...
  /intelligence/generate:
    post:
      description: "Generate comprehensive strategic intelligence reports from agent-collected\
        \ market data.\n    \n    This endpoint orchestrates the strategic intelligence\
        \ pipeline to:\n    1. Collect data from specified intelligence agents\n \
        \   2. Apply strategic analysis frameworks (SWOT, Porter's Forces)\n    3.\
        \ Generate actionable insights and recommendations\n    4. Create priority\
        \ communications for stakeholders\n    \n    **Strategic Analysis Capabilities:**\n\
        \    - Market trend analysis and competitive intelligence\n    - Risk assessment\
        \ and opportunity identification\n    - Strategic alignment scoring and recommendations\n\
        \    - Executive briefings and priority alerts\n    \n    **Agent Integration:**\n\
        \    - News monitoring and sentiment analysis\n    - Competitor tracking and\
        \ benchmarking\n    - Regulatory change detection\n    - Technology trend\
        \ analysis\n    \n    **Output Formats:**\n    - Strategic truths and organizational\
        \ insights\n    - Executive reports and briefings\n    - Priority communication\
        \ queue\n    - Strategic alignment scorecards"
      operationId: generate_strategic_intelligence_intelligence_generate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntelligenceRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntelligenceResponse'
          description: Generated strategic intelligence with insights, reports, and
            communications
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Generate Strategic Intelligence
      tags:
      - intelligence
...
  /intelligence/templates:
    get:
      description: Get available prompt templates.
      operationId: get_templates_intelligence_templates_get
      parameters:
      - in: query
        name: role
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
      - in: query
        name: category
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PromptTemplate'
                title: Response Get Templates Intelligence Templates Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Templates
      tags:
      - intelligence
    post:
      description: Create a new prompt template.
      operationId: create_template_intelligence_templates_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromptTemplate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptTemplate'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Template
      tags:
      - intelligence
  /intelligence/truths:
    get:
      description: Get strategic insights based on query criteria.
      operationId: get_strategic_insights_intelligence_truths_get
      parameters:
      - in: query
        name: categories
        required: false
        schema:
          anyOf:
          - items:
              $ref: '#/components/schemas/TruthCategory'
            type: array
          - type: 'null'
          title: Categories
      - in: query
        name: impact_levels
        required: false
        schema:
          anyOf:
          - items:
              $ref: '#/components/schemas/ImpactLevel'
            type: array
          - type: 'null'
          title: Impact Levels
      - in: query
        name: confidence_min
        required: false
        schema:
          anyOf:
          - maximum: 1.0
            minimum: 0.0
            type: number
          - type: 'null'
          title: Confidence Min
      - in: query
        name: date_from
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Date From
      - in: query
        name: date_to
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Date To
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 200
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/OrganizationalTruth'
                title: Response Get Strategic Insights Intelligence Truths Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Strategic Insights
      tags:
      - intelligence
...
servers:
- description: Production server
  url: https://api.livingtwin.com
- description: Staging server
  url: https://staging-api.livingtwin.com
- description: Development server
  url: http://localhost:8000
...